1) Проверка сертификатов ЭЦП при загрузке, сохранение результата в БД для отчета 2) Отчет о проверки ЭЦП, кнопка в заявлении для отдела приема и администратора 3) Выбор расчетного счета при создании счета из программы 4) Сокращенное имя автора в списках во всех документах,сокарщенное ФИО гл.эксперта в контрактах 5) Список регистраций колонка автор 6) Ошибка в исх.письме нашем: тип договорные документы, папка не была видна. 7) Согласование: при любом результате (согласовани/не согласовано) если закрыто, то доступ закрыт. Второй раз отправлять запрещено!!! Проверел 8) Контракт: при нажатии ОК срабатывает как отмена. Вопрос: сохранить? 9) Распорядительный акт: текст, 3000 знаков 10) Документ контракт: новые поля первичное заявление, номер первичного заявления 11) Рег.номер первичного контракта доступно для изменения в контракте 12) Перемещение файлов в другие папки после отправки исх.письма. Доступно администратору 13) В тему исх.письма добавлено название объекта во всех типах исх.писем. Соответственно объект переходит и в Согласование 14) Клиенту в исх.письме в прочих файлах можно крепить только 4 файла (2 с подписями)

+++ functions/file_upload.php
@@ -8,6 +8,9 @@ require_once(FRAME_WORK_PATH.'basic_classes/FieldSQLInt.php');
 require_once(ABSOLUTE_PATH.'controllers/Application_Controller.php');
 require_once(ABSOLUTE_PATH.'controllers/DocFlow_Controller.php');
 
+require_once(ABSOLUTE_PATH.'functions/PKIManager.php');
+require_once(ABSOLUTE_PATH.'functions/pki.php');
+
 include ABSOLUTE_PATH.'vendor/autoload.php';
  
 use Dilab\Network\SimpleRequest;
@@ -173,9 +176,26 @@ try{
 				$new_name = $resumable->uploadFolder.DIRECTORY_SEPARATOR.$_REQUEST['file_id'].($is_sig? '.sig':'');
 				rename($orig_file,$new_name);
 				chmod($new_name, 0664);
+				
+				//Если загружено все (файл + данные), делаем проверку подписи
+				if (
+				file_exists($file_doc = $resumable->uploadFolder.DIRECTORY_SEPARATOR.$_REQUEST['file_id'])
+				&&file_exists($file_doc_sig = $resumable->uploadFolder.DIRECTORY_SEPARATOR.$_REQUEST['file_id'].'.sig')
+				){
+					try{
+						$db_file_id = NULL;
+						FieldSQLString::formatForDb($dbLink,$_REQUEST['file_id'],$db_file_id);
+						
+						$pki_man = new PKIManager(PKI_PATH,PKI_CRL_VALIDITY,'error');
+						pki_log_sig_check($file_doc_sig, $file_doc, $db_file_id, $pki_man, $dbLink);
+					}
+					catch(Exception $e){
+					
+					}
+				}
 			}
 			catch(Exception $e){
-				unlink($orig_file);
+				if(file_exists($orig_file))unlink($orig_file);
 				throw $e;			
 			}
 		}
@@ -304,7 +324,7 @@ try{
 				chmod($new_name, 0664);			
 			}
 			catch(Exception $e){
-				unlink($orig_file);
+				if(file_exists($orig_file))unlink($orig_file);
 				throw $e;
 			}		
 		}
